benchfmt: new benchmark parsing package
This adds a new package for parsing the Go benchmark format. In
contrast with the storage/benchfmt package, which this deprecates, the
new package is a complete implementation of the format including
parsing file-level configuration, benchmark lines, and benchmark
names.

This is the beginning of a substantial redesign of the perf module,
which will culminate in a new benchstat that's substantially more
featureful and modular. These new packages can live next to the
existing packages, so it's not necessary to introduce a v2 of this
module.

The benchfmt API is designed for streaming use and expects the calling
package to transform the results into some other representation. One
such representation is coming in a following CL.

The reader API is also designed to process millions of records per
second on a typical laptop. This may seem unnecessary, but we've seen
data volumes like this in practice.

Change-Id: Ic5af0e936cf75da29372c2bb5747886f11ac7661